DIsable finalizers
I still don't have a firm grasp on the finalizers.
A big question I have is whether I can use the ArgoCD app
as the root app in the app of apps pattern.
I am leaning towards no right now,
because when I go to remove ArgoCD from the cluster
dependencies of the child apps are removed before
the child apps and that seems to prevent
the child apps from being removed.
Removing the finalizers for now.